Fix issues with Luma adjustment and HLG. Now it seems to be working. Interestingly enough, we have two different versions. The normal version uses existing code. Here Luma adjustment is done without accounting for system gamma. On the other hand, the second method tries to account for that as well. Surprisingly simple tests seem to say that these two are about the same in performance. Obviously the one that does not consider the system gamma is faster.
Also of note is that LUTs are not supported properly for HLG.
Although the forward TF seems to be okay (to go back to linear space)
the inverse is not and results in mismatches. Also the long LUT
method for HLG also seems to result in issues. Need to investigate
this further in the future.